Bug 23261: implement configuration portion of new Tor Launcher UI
Eliminate several wizard pages and include all of the bridge and
  proxy settings on one page.
Use the new Tor Browser logo.
Reuse Mozilla's help button from the Firefox hamburger menu and add
  help for the proxy settings..
In the wizard, use a "pageshow" event listener to show page-specific
  titles and to hide/show the navigation buttons as needed.
After the user chooses to restart the tor process, display the
  "Waiting for Tor to start" panel.
When tor is restarted by the user, disable networking so that
  bootstrapping does not begin right away; this allows the user to
  make configuration choices first.
Do not try to connect to the control port if tor is not running.