Bug 28039: Enable dump() if log method is 0
Without setting `browser.dom.window.dump.enabled` explicitly to `true`
it is not possible anymore to capture output in the terminal when the
log method is set to 0. We make sure that
`browser.dom.window.dump.enabled` is set to `true` in that case and
disable it again otherwise, unless Torbutton has enabled logging in a
terminal as well.

For context see:
https://bugzilla.mozilla.org/show_bug.cgi?id=379797
https://bugzilla.mozilla.org/show_bug.cgi?id=1395711